+++ R/chromatogramsOverlayGui.R
@@ -1,5 +1,4 @@
-
-
+## Function to create the GUI for the `chromatograms_overlay()` plot
 chrOverlayGui <- function (object){
     xrange <- round(unlist(rtime(object)), 2)
     yrange <- round(unlist(intensity(object)), 2)
@@ -7,12 +6,15 @@ chrOverlayGui <- function (object){
         column(
             width = 3,
             sliderInput("chrOverlay_xlim", "Resize X axes:",
-                        value = c(min(xrange), max(xrange)),
-                        min = min(xrange), max = max(xrange), dragRange = TRUE),
+                    value = c(min(xrange, na.rm = T), max(xrange, na.rm = T)),
+                    min = min(xrange, na.rm = T), max = max(xrange, na.rm = T),
+                    dragRange = TRUE),
             sliderInput("chrOverlay_ylim", "Resize Y axes:",
-                        value = c(min(yrange), max(yrange)),
-                        min = min(yrange), max = max(yrange), dragRange = TRUE),
-            colourInput("chrOverlay_color", "Select color: ", value = "#00000080"),
+                    value = c(min(yrange, na.rm = T), max(yrange, na.rm = T)),
+                    min = min(yrange, na.rm = T), max = max(yrange, na.rm = T),
+                    dragRange = TRUE),
+            colourInput("chrOverlay_color", "Select color: ",
+                        value = "#00000080"),
             numericInput("chrOverlay_bs", "Font size: ", value = 16, min = 1),
             numericInput("chrOverlay_pch", "Plotting symbol: ", value = 20,
                          min = 0, max = 25),