+++ R/zzz.R
@@ -1,6 +1,7 @@
 #' @importFrom utils packageVersion
 .onAttach <- function(libname, pkgname) {
   packageStartupMessage(
-      paste("\nThis is ChromatogramsVis version", packageVersion("ChromatogramsVis"), "\n",
+      paste("\nThis is ChromatogramsVis version",
+            packageVersion("ChromatogramsVis"), "\n",
             " See `?ChromatogramsVis` to get started.\n"))
 }