+++ NEWS.md
@@ -1,4 +1,10 @@
 # ChromatogramsVis 0.1
 
+## Changes in 0.1.2
+
+- Fix bug data.frame visualization.
+- Refactor and clean.
+
 ## Changes in 0.1.1
 
+- Initial version.